F_T32247 move ReportBundle Value Objects to core
- adjust use statements
