+++ demosplan/DemosPlanCoreBundle/Logic/Report/ReportMessageConverter.php
@@ -14,10 +14,10 @@ use DemosEurope\DemosplanAddon\Contracts\Config\GlobalConfigInterface;
 use demosplan\DemosPlanCoreBundle\Entity\Report\ReportEntry;
 use demosplan\DemosPlanCoreBundle\Permissions\PermissionsInterface;
 use demosplan\DemosPlanCoreBundle\Twig\Extension\DateExtension;
-use demosplan\DemosPlanReportBundle\ValueObject\ProcedureFinalMailReportEntryData;
-use demosplan\DemosPlanReportBundle\ValueObject\RegisteredInvitationReportEntryData;
-use demosplan\DemosPlanReportBundle\ValueObject\StatementFinalMailReportEntryData;
-use demosplan\DemosPlanReportBundle\ValueObject\UnregisteredInvitationReportEntryData;
+use demosplan\DemosPlanCoreBundle\ValueObject\Report\ProcedureFinalMailReportEntryData;
+use demosplan\DemosPlanCoreBundle\ValueObject\Report\RegisteredInvitationReportEntryData;
+use demosplan\DemosPlanCoreBundle\ValueObject\Report\StatementFinalMailReportEntryData;
+use demosplan\DemosPlanCoreBundle\ValueObject\Report\UnregisteredInvitationReportEntryData;
 use Exception;
 use Psr\Log\LoggerInterface;
 use Seld\JsonLint\JsonParser;
